Convert existing DropDowns to normal panels
